Move get_latest_instance_version_from_json function to _utils.py and update configure_minimum_instance_version to use it
